feat: implement of additional functionality
